+++ common/src/main/java/org/ayamemc/paperdollrender/mixin/hook/MinecraftClientMixin.java
