+++ common/src/main/java/org/ayamemc/paperdollrender/mixin/patch/ImmediateMixin.java
