+++ common/src/main/java/org/ayamemc/paperdollrender/mixin/retexture/PressableWidgetMixin.java
@@ -1,0 +1,42 @@
+/*
+ *     Highly configurable paper doll mod, well integrated with Ayame.
+ *     Copyright (C) 2024  LucunJi(Original author), CrystalNeko, HappyRespawnanchor, pertaz(Icon Designer)
+ *
+ *     This file is part of PaperDoll Render.
+ *
+ *     PaperDoll Render is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU Lesser General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     PaperDoll Render is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU Lesser General Public License for more details.
+ *
+ *     You should have received a copy of the GNU Lesser General Public License
+ *     along with PaperDoll Render.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package org.ayamemc.paperdollrender.mixin.retexture;
+
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.gui.components.AbstractButton;
+import net.minecraft.resources.ResourceLocation;
+import org.ayamemc.paperdollrender.config.view.Retextured;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(AbstractButton.class)
+public class PressableWidgetMixin {
+    @WrapOperation(method = "renderWidget", at = {
+            @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/GuiGraphics;blitSprite(Lnet/minecraft/resources/ResourceLocation;IIII)V")
+    })
+    public void drawTransparentTextFieldTexture(GuiGraphics instance, ResourceLocation texture, int x, int y, int width, int height, Operation<Void> original) {
+        if (this instanceof Retextured retextured)
+            original.call(instance, retextured.retexture(texture), x, y, width, height);
+        else original.call(instance, texture, x, y, width, height);
+    }
+}
