--- fabric/src/main/java/org/ayamemc/paperdollrender/mixin/retexture/TextFieldWidgetMixin.java
@@ -1,55 +1,0 @@
-/*
- *     Highly configurable paper doll mod, well integrated with Ayame.
- *     Copyright (C) 2024  LucunJi(Original author), CrystalNeko, HappyRespawnanchor
- *
- *     This file is part of PaperDoll Render.
- *
- *     PaperDoll Render is free software: you can redistribute it and/or modify
- *     it under the terms of the GNU Lesser General Public License as published by
- *     the Free Software Foundation, either version 3 of the License, or
- *     (at your option) any later version.
- *
- *     PaperDoll Render is distributed in the hope that it will be useful,
- *     but WITHOUT ANY WARRANTY; without even the implied warranty of
- *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *     GNU Lesser General Public License for more details.
- *
- *     You should have received a copy of the GNU Lesser General Public License
- *     along with PaperDoll Render.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-package org.ayamemc.paperdollrender.mixin.retexture;
-
-import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
-import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
-import com.mojang.blaze3d.systems.RenderSystem;
-import net.minecraft.client.gui.GuiGraphics;
-import net.minecraft.client.gui.components.AbstractWidget;
-import net.minecraft.client.gui.components.EditBox;
-import net.minecraft.network.chat.Component;
-import net.minecraft.resources.ResourceLocation;
-import org.ayamemc.paperdollrender.api.config.view.Retextured;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-
-@Mixin(EditBox.class)
-public abstract class TextFieldWidgetMixin extends AbstractWidget {
-    public TextFieldWidgetMixin(int x, int y, int width, int height, Component message) {
-        super(x, y, width, height, message);
-    }
-
-    @WrapOperation(method = "renderWidget", at = {
-            @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/GuiGraphics;blitSprite(Lnet/minecraft/resources/ResourceLocation;IIII)V")
-    })
-    public void drawTransparentTextFieldTexture(GuiGraphics instance, ResourceLocation texture, int x, int y, int width, int height, Operation<Void> original) {
-        if (this instanceof Retextured retextured) {
-            instance.setColor(1.0f, 1.0f, 1.0f, this.alpha);
-            RenderSystem.enableBlend();
-            RenderSystem.enableDepthTest();
-            original.call(instance, retextured.retexture(texture), x, y, width, height);
-            instance.setColor(1.0f, 1.0f, 1.0f, 1.0f);
-            return;
-        }
-        original.call(instance, texture, x, y, width, height);
-    }
-}
